% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heston.R
\name{HESTON}
\alias{HESTON}
\title{Title}
\usage{
HESTON(rho, sigma, lambda, t, r, theta, V0, x0, n_steps, paths)
}
\arguments{
\item{rho}{Correlation for the Brownian motions}

\item{sigma}{Variance parameter for the Vt process}

\item{lambda}{Mean-reversion rate parameter}

\item{t}{Amount of periods to simulate}

\item{r}{Interest rate}

\item{theta}{Long term mean for Vt}

\item{V0}{Value of Vt at the start}

\item{x0}{Initial log of stock price}

\item{n_steps}{Amount of time steps}

\item{paths}{Amount of paths to simulate}
}
\value{
HestonMc object where every row is a simulation and every column a timestep
}
\description{
Function that simulates a Heston model using Monte-Carlo
}
\examples{
 HESTON(rho=-0.5, sigma=0.1, lambda=2, t=10, r=0.02, 
 theta=0.1^2, V0=0.1^2, x0=log(100), n_steps=5, paths=10)

}
