% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heston.R
\name{HESTON_mc}
\alias{HESTON_mc}
\title{Title}
\usage{
HESTON_mc(prices, volatility)
}
\arguments{
\item{prices}{List of price paths}

\item{volatility}{List of volatility}
}
\value{
Returns HESTON_mc object
}
\description{
HESTON_mc constructor
}
