\name{KWAENFInR-package}
\alias{KWAENFInR-package}
\alias{KWAENFInR}
\docType{package}
\title{
  Efficient implementations of mathematical finance models 
}
\description{
 This package provides an efficient implementation of numerical methods used in financial option pricing. Specifically the Heston model for pricing stock options and LIBOR market model for interest rate derivatives. 
}

\author{
Quinn Ramakers.

Maintainer: Quinn Ramakers <q.ramakers@student.maastrichtuniversity.nl>
}


