% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIBOR.R
\name{LIBORmc}
\alias{LIBORmc}
\title{Title}
\usage{
LIBORmc(rates, DF, drift)
}
\arguments{
\item{rates}{list of forward rates}

\item{DF}{list of discount factors}

\item{drift}{list of drifts}
}
\value{
object of libormc class
}
\description{
S3 class constructor
}
