% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIBOR.R
\name{mcLIBOR}
\alias{mcLIBOR}
\title{Title}
\usage{
mcLIBOR(forwardrates, dT, sigma, n)
}
\arguments{
\item{forwardrates}{Vector of forward rates with the same frequency.}

\item{dT}{Timesteps where 1 is equal to a year. This should be equal to the time between forward rates.}

\item{sigma}{Volatility matrix, in case of constant volatility provide a matrix with only a singular value for all indexes.}

\item{n}{Number of simulations}
}
\value{
Object of class LIBORmc containing all matrixes simulated.
}
\description{
This function simulates LIBOR market rate models and returns an object with 
the matrixes for multiple simulations to be used for pricing of desired derivative.
}
\details{
We implement the LIBOR market rate model based on martingale pricing and compensate 
using the Radon-Nikodyn derivative to add a drift terms as found in the drift matrix. 
This method always simulates untill the highest maturity is reached.
}
